Re-Enable HTTPS-only Cookies
Modify the backend to require HTTPS connections for sending
authentication cookies. This means we can no longer hit the admin or
login routes directly, since the server runs on HTTP, we must do so
through an HTTPS proxy like nginx in production or the webpack dev
server proxy in development.