Add Development Request Logging, Restrict SQL Logging
Add an `Environment` type to the backend that is parsed from the
`ENVIRONMENT` env var when building the `Config` type. This can be used
to switch on functionality based on whether the server is running in
development or production.

Add a request logger to the server that only operates when run under the
`Development` environment. Modify SQL query logging so it only logs
during `Development` as well.