Fix Initial Loading of Homepage For Every Visit
Fix a bug that caused the Home route to be loaded(along with it's API
data) whenever the client was opened - even if it was opened to a
different route. When opening a URL to a post's page, the app would load
the post list data, start rendering it, then change the route to the
post details page.

We fix this by pulling the initial state before launching the Halogen
app & passing it in as input to the `BaseSite` component.