Render Markdown in Post Content, Add Syntax Highlighting
Use the markdown-it & html-parser-halogen packages to parse the markdown
in post bodies into halogen HTML elements & render them on the View Blog
Post page. Add highlight.js via the FFI to add highlighting classes &
CSS to fenced code blocks with a specified language.

Add a new Markdown effect typeclass for rendering markdown strings into
HTML strings via `markdown-it`. Keep a renderer instance in the AppEnv
for re-use but add an unsafe rendering function that has a pure
signature for emergencies.