Unify Api & JSON Errors on Frontend
Introduce a new ApiError type to the Api module for combining JSON
decoding errors with HTTP request errors & allowing views to render both
in a single type instead of already-stringified versions of either
error.

Introduce a new StatusCodeError error value for when the server responds
with a Status Code of 400 or above. These indicate client or server
errors. To make this possible we have to set our request's response
format from Json to String. This is necessary because the server does
not enforce JSON response bodies when an error occurs, which causes body
parsing of Json response formats to pre-emptively fail with a
ResponseBodyError. By doing the JSON parsing & decoding ourselves, we
can do a status code check _before_ looking at the response body.