Add BlogCategory Table & Include Category in Responses
Add a new blog_category table to the database along with a link in the
blog_post table for the BlogCategory a post is in.

Fetch the Category when fetching a post & include it in the List &
Details responses. Take a BlogCategoryId when creating a new Post.

Add a new Categories section to the Sidebar response data, with a list
of categories & the number of posts within each.