Add Category Pages, Sidebar Sectin, & Post Meta
Add a new page to the website for showing all posts within a certain
Category.

Add a new Categories section to the sidebar with a list of categories &
their post counts, linking to the new Category page.

Add the Category link to the Post Meta section of a post, in both the
List & Details views.