Add Logout Route, Refactor Authentication
Add a new `userLogout` route that removes the cookies set by the
`userLogin` route.

Rename the JWTToken effect monad into AuthToken & add a method for
fetching the CookieSettings. Set a MaxAge for the cookies to persist
logins for 1 year.

Don't include the AuthResult in the response bodies when throwing 403s
from authenticated routes.