Add Authorization, Login Page, & Admin Dashboard to Client
Add a login page to the website which collects a username & password &
attempts to log in to the API. Display an invalid credentials message on
401 errors, generic message on other Api Errors, & redirect to the Admin
Dashboard on success. An optional `redirectTo` query parameter is
supported. On success, we will redirect to the given internal URL
instead.

Add an Admin Dashboard page which is a simple placeholder for validating
the current authorization status & providing a way to log out.

Add a new Auth effect monad which functionality to toggle & query the
authorization status of the user. Auth status is persisted in the AppEnv
and local storage. On bootup, the app attempts to pull an initial
AuthStatus from local storage, defaulting to Anonymous.

Modify the UpdateRoute query of the BaseSite component. When passed an
admin route & not logged in, redirect to the Login page w/ a redirect
back to the desired admin route. When passed the login page while logged
in, redirect to the route in the query parameter, defaulting to the
Dashboard. Maintain existing behavior for all other situations.