Add Support for Serving & Manipulating Uploaded Media
Add a new `cfgMediaDirectory` field to the `Config` type, represnting
the folder where file uploads are stored. Add a `Media` typeclass for
filesystem manipulations within this folder.

Add new API routes for listing the contents of a media directory,
creating new folders, & uploading files.

Modify the server so that it serves the media directory when the
`Environment` is set to `Development`.

Add testing infrastrucure that allows us to test code that uses the
`Media` typeclass via an in-memory filesystem, instead of requiring an
actual media directory to manipulate.