Add a Media Management Page to Site Admin
Add a new `AdminMediaList` page & component to the admin site for
managing uploaded files. This page includes a listing of the
upload/media directory, navigation through sub-directories, & the
ability to create sub-directories & upload files.

Add a new `FileUpload` typeclass for converting the contents of uploaded
`File` objects into their base64 encodings for use with the API.

Modify the dev server so it proxies requests to the `/media` path to the
backend server. This assumes the backend is running in development mode.
In production, this mapping should be handled by the webserver.