Clear Input & Display Feedback After File Uploads
Modify the `AdminMediaList` Page so that it clears the file input after
a successful upload & displays either an error or a link to the uploaded
file.

Add a `clearInputValue` function to the `FileUpload` typeclass, which
calls the FFI to clear an `Element`'s `value` property. For file inputs,
this removes the file & stops rendering it's name.