Add Link DB Model & Public API Routes
Add new `Link` & `LinkCategory` database models as well as API routes
for redirecting to a links location, fetching a mapping of all
categories to their child categories & direct links, as well as fetching
the mapping for a specific category.

Begin documenting helpful queries when porting data from the old
website's database in `SITE_MIGRATION.md`.

Modify the `dropAllTables` helper in the `Test.Setup` module so that it
ignores `ForeignRef`s to itself - the asummption was these would be
handled by `SelfReference` case, but that does not appear to be the
case for the `LinkCategory` model.