Add Link Pages & Proxy the API Redirect
Add new `LinkView` & `LinkViewCategory` page components, Routes, API
requests, and associated types for rendering tables of LinkCategories &
their Links - for all categories and for a specific category.

Add a new `openInNewTab` method to the `Navigation` typeclass for
opening an arbitrary URL in a new browser tab using the `open()`
function on the `window` object.

Modify the `BaseSite` component's `renderHeader` function so we mark a
top-level nav item as active if we are on any of it's child pages. Blog
pages mark `Home` as active, Link pages mark `Link` as active.

Modify the webpack configuration to relay/proxy the redirection requests
to the API server's 302 link redirect route during development. In
production, this should be handled in the nginx server's configuration.