Cache Data For Blog Sidebars
The worst performing code in our Blog API requests is generating the
sidebar data. This commit improves performance by implementing a caching
interface that stores the sidebar data for the frontend routes & allows
the admin routes to bust the cache when a post changes.

Add a new `Caches` module for aggregating all the cached data as well as
functions for initializing the entire cache & regenerating specific
parts of the cache. Move the `BlogSidebarData` generation code from the
`BlogPosts` Handler module to the `Caches.Blog` module.

Add the cache to the global application state, wrapping it in a `TVar`
for safe modification across requests. Implement a new `Caches`
typeclass for fetching & busting the cache. Use the new `Caches`
typeclass constraint in the `BlogPosts` & `Admin` handler modules.

Add support for the `Caches` typeclass to the test runner. Refactor the
`customTestRunner` to return a new `CustomTestResult` type instead of
ever-increasing tuples. This new type allows inspection of the test's
result, the final media filesystem, & the final cache state.