Delay Rendering of Loading State When Changing Pages
Page navigation has been abrupt - when you change pages, the rendered
page immediately jumps to the loading state & then jumps again when the
page data is finished loading. Normally this would be a reasonable flow,
but with our quick API responses, this means the loading state only
flashes on screen for dozens of milliseconds, which means two quick page
changes within 50ms. This is jarring.

This commit fixes this abruptness by delaying rendered page changes
until 250ms has passed or the page data's API request has completed -
whichever occurs first.

Add new typeclasses to the App module, one for retrieving a subscription
to page data changes, one for listening to page data changes, and
another for sending page data loaded notifiations to the subscription.

Add a `mkPageDataNotifierEval` wrapper function around Halogen's
`mkEval` function. This requires the component state to have an
`apiData` field & watches the field for a Loading->Loaded state
transition & fires off the `pageDataReceived` effect when that
transition occurs.

Modify the BaseSite & BaseAdmin components to optionally store the
previous page. When rendering these components, show the previous page
if defined along with a hidden verison of the current page. This allows
the current page to make it's initialization API requests while the user
still sees the previous page.

Modify the BaseSite component so that it sets the previous page state
field when navigating to a new page & starts an async thread that clears
the field after 250ms. The component also listens for events from the
Page Data subscription and clears the previous page state & kills the
async thread when an event is received. This, combined with the new
rendering functionality, is what delays display of new pages and reduces
the jitter.

Modify all the Pages components so they either utilize the new
`mkPageDataNotifierEval` function(if they have an `apiData` field) or
manually call the `pageDataReceived` effect on initialization(if they do
not have any data to fetch before they can fully render).