Dynamically Set Title & Meta Description on Page Load
Modify the page data subscription so that we can set the title tag &
description meta tags after the current page's data has finished
loading.

Add a new `SEOData` type to the `App` module for storing a page's title
& description. Modify the Page Data Subscription so the produced events
contain an optional SEOData value. This will contain a value when the
data is loaded & have no value when the loading timeout fires.

Modify the `mkPageDataNotifierEval` function so it takes a function to
build the SEOData from a component's state & API response. Send this
data on a successful response, with standardized SEOData for error
responses.

Add a new `MetaElements` type to the `App` module for storing the HTML
document & meta elements we need for updating the meta description. Add
a `MetaTags` typeclass that uses the `MetaElements` in the `AppEnv` to
update the page title & description from an `SEOData` value.

Modify the `BaseSite` component's `handleAction` function so that it
set's the meta tags when it's Page Data Subscription handler action
receives an SEOData value.

Update all page components so they pass an `SEOData` value to the
`pageDataReceived` call or an `SEOData`-generating function to the
`mkPageDataNotifierEval` fucntion.

Update the webpack configuration to add the initial meta description
elements in the generated `index.html`.