Add Initial Docker Compose Deployment Configuration
Add some initial Dockerfiles, an Nginx Config, a Docker Compose config,
& a Makefile for building backend & frontend servers and orchestrating
deployment of the backend, nginx, & postgresql services via Docker
Compose. This work is all still in-progress, we are just ready to test
the CI flow.

Add a new CI action that builds the docker images, tags them with
`latest` & the commit hash, and pushes them to Dockerhub. We will rework
this so a separate job is responsible for tagging to latest for only the
`master` branch.

We need to integrate secrets management into this - as currently the
database is password-less & the JWK is ephemeral. This is okay as this
code won't actually go into production until that is integrated.