Docker Health Checks & Tweaks
Modify the nginx & api Dockerfiles so they include `HEALTHCHECK`
commands to ensure the services are healthy. Modify the docker-compose
file to include a custom healthcheck for the database & modify the
`depends_on` keys so that dependencies are healthy before the dependent
service is started.

Remove the `expose` ports for the internal services as they are not
necessary.

Add the `tty` flag for the api service as that is required for it to
print to stdout.

Re-order some commands in the Dockerfiles to improve build caching.