Add Prerender Container & Switch Nginx to HTTPS
Add a new `prerender` container to the docker compose configuration as
well as a `/deploy/prerender` folder with a small npm project for
running a prerender server & a Dockerfile for building the container.

Switch Nginx from the 8080 http port to the classic 80/443 port & enable
SSL on the 443 port server. Use HTST headers to redirect clients to the
secure server. Modify the nginx Dockerfile to create a temporary
self-signed certificate - this will be used until certbot overrides it.

Modify the nginx configuration to redirect bot user-agents to the
prerender server, which will loop back to the nginx server & return the
fully rendered content to the bot, instead of just the scaffolding from
the `index.html` file.

Add long-term caching for our compiled static files, like the JS, CSS, &
favicons.