Add SSL Stapling, Fix Hostname Errors in Nginx Proxy
Turn on ssl stapling verification in nginx.

Modify the nginx proxy_pass directives so they use variables from `set`
directives instead of using hostnames in the actual `proxy_pass` lines.
Putting a hostname in proxy_pass causes a DNS lookup when nginx starts -
which may be too early for the system to resolve our internal DNS
hostnames like `api` or `prerender`. Putting these in a variable allows
the hostname lookup to wait until the proxy_pass is actually executed.

This should prevent the `host not found in upstream` errors we
occasionally see when bringing the whole system up. It should also allow
us to drop the container dependencies for the nginx container, which
means we could roll out api-only changes without causing the nginx
container to reboot as well.