Use HTTPS For Nginx Health Check
Modify the nginx Dockerfile to use HTTPS for the healthcheck command.
This also requires setting curl's `insecure` flag since we have
self-signed certificates. This shouldn't be a problem since we talk
directly to localhost instead of othe domain & will stop the healthcheck
from logging 302 responses due to the HTTPS redirect.