Use Certbot for Production Certificates
Add a new `certbot` service to our Docker Compose configuration. At the
moment, this simply uses the upstream docker image. We use one volume to
store the generated ceritficates for nginx and another volume for the
verification challenges that nginx must serve. A bootstrap docker
compose override configuration is added with the certbot command to
create the initial certificates.

Modify the nginx Dockerfile & config file to put the self-signed
certificates in the same location that certbot will write them to. This
allows us bootstrap the https server with self-signed certs and then
reload it after certbot has made proper certificates.

We need to figure out renewals - should certbot actually be folded into
the nginx service? Should we have an entrypoint script that loops
between calling the renewal command & sleeping for 24hrs? The latter
would require a daily cronjob to reload the nginx server.

All of this was roughly tested in a production enviornment, but we'll
wait for the images to be built & put through the normal deployment
process before considering this a success & documenting the initial &
continuous deployment flows.