Documentation & Tweaks for Initial Deployment
This gets us 95% of the way to considering the initial deployment as
complete!

Add a plethora of `remote-*` commands to the `Makefile` for managing the
production docker system from our local machines.

Add documentation on running through the initial deployment to the
`README`.

Add a script for migrating data from a backup of the old database, add a
`Makefile` command for copying the script & data to the database,
running the data migration, and cleaning up.

Modify the nginx config so it redirects HTTP requests to the bare
`sleepanarchy.com` domain.

Increase the start period & timeout lengths for the prerender
container's healthcheck command.

Modify the bootstrap docker compose file to migrate the database schema
& to remove the self-signed certificates before running the certbot
initalization.

Litter some TODOs around the codebase, these should be the final bits
before we consider ourselves 100% deployed & stable.