Set Meta Header for Proper Prerender Status Codes
Expand the page data subscription with a new `PageLoadData` type that
nests the pre-existing `SEOData` and adds a `apiStatusCode` field as
it's sibling. Use this new field to set the `prerender-status-code` meta
element's content when we receive a response from the API. This meta
element tells prerender what status code it should return with the
rendered response.

For successful responses, we always tell prerender to return a 200. For
status code errors, we forward the status code from the API server. For
generic HTTP errors, we throw 500s. For JSON decoding errors we throw
429s.