Add Healthcheck Route & Custom Production Logger
Add a new `/healthcheck` route that queries the database & cache by
re-using the `getBlogPosts` handler but returns no content.

Customize the production logger so that it pulls client IPs from the
`X-Real-IP` header and ignore logging of the healthcheck route.

Modify the nginx configuration so that it passes the `X-Real-IP` header
when proxying requests to the API server.